[#noissue] Add IOUtils  - move to common module
