+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/util/JarReader.java
@@ -17,11 +17,11 @@
 
 package com.navercorp.pinpoint.profiler.util;
 
+import com.navercorp.pinpoint.common.util.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.ByteArrayOutputStream;
-import java.io.Closeable;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -108,30 +108,15 @@ public class JarReader {
                 logger.warn("jarFile read error jarFile:{} jarEntry{} {}", jarFile, jarEntry, ioe.getMessage(), ioe);
                 throw ioe;
             } finally {
-                close(inputStream);
+                IOUtils.closeQuietly(inputStream);
             }
         }
 
         public byte[] read(InputStream input) throws IOException {
             this.output.reset();
-            read(input, output);
+            IOUtils.copy(input, output, buffer);
             return output.toByteArray();
         }
 
-        public void read(InputStream input, OutputStream output) throws IOException {
-            int readIndex;
-            while ((readIndex = input.read(buffer)) != -1) {
-                output.write(buffer, 0, readIndex);
-            }
-        }
-
-        private void close(Closeable closeable) {
-            if (closeable != null) {
-                try {
-                    closeable.close();
-                } catch (IOException ignore) {
-                }
-            }
-        }
     }
 }
